Precise slider control

+++ wrapper/src/main/java/io/github/solclient/client/ui/component/Component.java
@@ -235,7 +235,8 @@ public abstract class Component extends NanoVGManager {
 
 	public boolean mouseClickedAnywhere(ComponentRenderInfo info, int button, boolean inside, boolean processed) {
 		if (dialog != null) {
-			boolean insideDialog = dialog.getBounds().contains(info.getRelativeMouseX(), info.getRelativeMouseY());
+			boolean insideDialog = dialog.getBounds().contains((int) info.getRelativeMouseX(),
+					(int) info.getRelativeMouseY());
 
 			if (dialog.mouseClickedAnywhere(transform(info, dialog.getBounds()), button, insideDialog, processed))
 				processed = true;
@@ -256,7 +257,8 @@ public abstract class Component extends NanoVGManager {
 				Rectangle bounds = getBounds(component);
 
 				if (component.mouseClickedAnywhere(transform(info, bounds), button,
-						inside && bounds.contains(info.getRelativeMouseX(), info.getRelativeMouseY()), processed))
+						inside && bounds.contains((int) info.getRelativeMouseX(), (int) info.getRelativeMouseY()),
+						processed))
 					return true;
 			}
 		} catch (ConcurrentModificationException error) {
@@ -283,7 +285,7 @@ public abstract class Component extends NanoVGManager {
 
 	public boolean mouseReleasedAnywhere(ComponentRenderInfo info, int button, boolean inside) {
 		if (dialog != null && dialog.mouseReleasedAnywhere(transform(info, dialog.getBounds()), button,
-				dialog.getBounds().contains(info.getRelativeMouseX(), info.getRelativeMouseY())))
+				dialog.getBounds().contains((int) info.getRelativeMouseX(), (int) info.getRelativeMouseY())))
 			return true;
 
 		if (onReleaseAnywhere != null && onReleaseAnywhere.onClick(info, button))
@@ -296,7 +298,7 @@ public abstract class Component extends NanoVGManager {
 			Rectangle bounds = getBounds(component);
 
 			if (component.mouseReleasedAnywhere(transform(info, bounds), button,
-					bounds.contains(info.getRelativeMouseX(), info.getRelativeMouseY())))
+					bounds.contains((int) info.getRelativeMouseX(), (int) info.getRelativeMouseY())))
 				return true;
 		}
 
@@ -321,7 +323,7 @@ public abstract class Component extends NanoVGManager {
 		for (Component component : subComponents) {
 			Rectangle bounds = getBounds(component);
 
-			if (!bounds.contains(info.getRelativeMouseX(), info.getRelativeMouseY()))
+			if (!bounds.contains((int) info.getRelativeMouseX(), (int) info.getRelativeMouseY()))
 				continue;
 
 			if (action.test(component, transform(info, bounds)))
