+++ wrapper/src/main/java/io/github/solclient/client/ui/component/ComponentRenderInfo.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class ComponentRenderInfo {
 
-	private final int relativeMouseX, relativeMouseY;
+	private final float relativeMouseX, relativeMouseY;
 	private final float partialTicks;
 
 }
