+++ wrapper/src/main/java/io/github/solclient/client/ui/screen/mods/ColourPickerDialog.java
@@ -95,7 +95,7 @@ public class ColourPickerDialog extends BlockComponent {
 
 		if (selectedSlider != -1) {
 			colour = colour.withComponent(selectedSlider,
-					MathHelper.clamp(info.getRelativeMouseX() - RGB_OFFSET_LEFT, 0, 255));
+					(int) MathHelper.clamp(info.getRelativeMouseX() - RGB_OFFSET_LEFT, 0, 255));
 			callback.accept(colour);
 			updateHex();
 		}
@@ -214,7 +214,7 @@ public class ColourPickerDialog extends BlockComponent {
 		for (int component = 0; component < 4; component++) {
 			Rectangle rectangle = new Rectangle(RGB_OFFSET_LEFT, RGB_OFFSET_TOP + component * RGB_SPACING, 256, 11);
 
-			if (rectangle.contains(info.getRelativeMouseX(), info.getRelativeMouseY()))
+			if (rectangle.contains((int) info.getRelativeMouseX(), (int) info.getRelativeMouseY()))
 				return component;
 		}
